Add the first OpenTag CLI
