+++ packages/cli/tsconfig.json
@@ -6,7 +6,9 @@
   },
   "include": ["src/**/*.ts"],
   "references": [
-    { "path": "../../packages/dispatcher" },
-    { "path": "../../packages/local-runtime" }
+    { "path": "../client" },
+    { "path": "../core" },
+    { "path": "../lark" },
+    { "path": "../local-runtime" }
   ]
 }
