+++ packages/lark/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./normalize.js";
+export * from "./inbound.js";
+export * from "./ingress.js";
 export * from "./outbound.js";
 export * from "./render.js";
