+++ tsconfig.json
@@ -5,10 +5,13 @@
     { "path": "packages/client" },
     { "path": "packages/dispatcher" },
     { "path": "packages/github" },
+    { "path": "packages/lark" },
+    { "path": "packages/local-runtime" },
     { "path": "packages/slack" },
     { "path": "packages/telegram" },
     { "path": "packages/runner" },
     { "path": "packages/store" },
+    { "path": "packages/cli" },
     { "path": "apps/dispatcher" },
     { "path": "apps/github-probot" },
     { "path": "apps/lark-events" },
